Removed unneeded dependencies, ground texture painting now applies pathing, small Qol.

+++ HiveWE/PathingMap.h
@@ -2,8 +2,8 @@
 
 class PathingMap {
 	public:
-	size_t width;
-	size_t height;
+	int width;
+	int height;
 
 	GLuint pathing_texture;
 	std::vector<uint8_t> pathing_cells;
